// src/generated/models/access-control-environment-assignment-model.ts
import type { AccessControlEntry } from './access-control-entry-model';
import type { Environment } from './environment-model';

export interface AccessControlEnvironmentAssignment {
  /**
   * @displayName Access Control Environment Assignment
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Environment Assignment Name
   * @description Human-readable record identifier
   */
  environmentAssignmentName: string;
  /**
   * @displayName Access Control Entry
   * @validationRule Required for create/update operations
   */
  accessControlEntry: Pick<AccessControlEntry, 'id' | 'accessControlEntryName'>;
  /**
   * @displayName Environment
   * @validationRule Required for create/update operations
   */
  environment: Pick<Environment, 'id' | 'environmentName'>;
}

export const _AccessControlEnvironmentAssignment = 'AccessControlEnvironmentAssignment' as const;